Finish seeders

+++ OutpostImmobile.Persistence/Domain/Routes/LocationMarkerEntity.cs
@@ -4,8 +4,8 @@ public class LocationMarkerEntity
 {
     public Guid Id { get; set; }
 
-    public required int Longitude { get; set; }
-    public required int Latitude { get; set; }
+    public required double Longitude { get; set; }
+    public required double Latitude { get; set; }
 
     public ICollection<RouteEntity> Routes { get; set; }
 }
